# docs/swagger.yaml
basePath: /
definitions:
...
  dtos.InvoiceItemOutput:
    properties:
      category_id:
        example: 660e8400-e29b-41d4-a716-446655440001
        type: string
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      description:
        example: iPhone 16 Pro
        type: string
      id:
        example: 880e8400-e29b-41d4-a716-446655440003
        type: string
      installment_amount:
        description: Valor desta parcela
        example: "833.25"
        type: string
      installment_label:
        description: 'Ex: "3/12" ou "À vista"'
        example: 3/12
        type: string
      installment_number:
        description: 1 a N
        example: 3
        type: integer
      installment_total:
        description: Total de parcelas
        example: 12
        type: integer
      invoice_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      purchase_date:
        description: YYYY-MM-DD
        example: "2025-01-15"
        type: string
      total_amount:
        description: Valor total da compra original
        example: "9999.00"
        type: string
      updated_at:
        example: "2025-01-20T08:00:00Z"
        type: string
    type: object
  dtos.InvoiceListOutput:
    properties:
      card_id:
        example: 770e8400-e29b-41d4-a716-446655440002
        type: string
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      currency:
        enum:
        - BRL
        - USD
        - EUR
        example: BRL
        type: string
      due_date:
        example: "2025-01-10"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      item_count:
        example: 12
        type: integer
      reference_month:
        example: 2025-01
        type: string
      total_amount:
        example: "9999.00"
        type: string
    type: object
  dtos.InvoiceOutput:
    properties:
      card_id:
        example: 770e8400-e29b-41d4-a716-446655440002
        type: string
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      currency:
        enum:
        - BRL
        - USD
        - EUR
        example: BRL
        type: string
      due_date:
        description: YYYY-MM-DD
        example: "2025-01-10"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      item_count:
        example: 12
        type: integer
      items:
        items:
          $ref: '#/definitions/dtos.InvoiceItemOutput'
        type: array
      reference_month:
        description: YYYY-MM
        example: 2025-01
        type: string
      total_amount:
        example: "9999.00"
        type: string
      updated_at:
        example: "2025-01-20T08:00:00Z"
        type: string
      user_id:
        example: 660e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  dtos.InvoicePaginatedOutput:
    properties:
      data:
        items:
          $ref: '#/definitions/dtos.InvoiceListOutput'
        type: array
      pagination:
        $ref: '#/definitions/dtos.InvoicePaginationMeta'
    type: object
  dtos.InvoicePaginationMeta:
    properties:
      has_next:
        example: true
        type: boolean
      limit:
        example: 20
        type: integer
      next_cursor:
        example: eyJmIjp7InJlZmVyZW5jZV9tb250aCI6IjIwMjUtMDEifX0
        type: string
    type: object
...
  dtos.PurchaseCreateInput:
    properties:
      card_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      category_id:
        example: 660e8400-e29b-41d4-a716-446655440001
        type: string
      currency:
        description: ISO 4217 (e.g., "BRL")
        enum:
        - BRL
        - USD
        - EUR
        example: BRL
        type: string
      description:
        example: iPhone 16 Pro
        type: string
      installment_total:
        description: 1 para à vista
        example: 12
        maximum: 48
        minimum: 1
        type: integer
      purchase_date:
        description: YYYY-MM-DD
        example: "2025-01-15"
        type: string
      total_amount:
        description: String decimal (e.g., "1200.00")
        example: "9999.00"
        type: string
    type: object
  dtos.PurchaseCreateOutput:
    properties:
      items:
        items:
          $ref: '#/definitions/dtos.InvoiceItemOutput'
        type: array
    type: object
  dtos.PurchaseUpdateInput:
    properties:
      category_id:
        example: 660e8400-e29b-41d4-a716-446655440001
        type: string
      description:
        example: iPhone 16 Pro Max
        type: string
      total_amount:
        example: "10999.00"
        type: string
    type: object
  dtos.PurchaseUpdateOutput:
    properties:
      items:
        items:
          $ref: '#/definitions/dtos.InvoiceItemOutput'
        type: array
    type: object
...
  httperrors.ProblemDetail:
    properties:
      detail:
        type: string
      errors:
        additionalProperties: {}
        type: object
      instance:
        type: string
      request_id:
        type: string
      status:
        type: integer
      timestamp:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Jailton Junior
    url: https://github.com/jailtonjunior94
  description: |-
    API para gestão financeira pessoal com suporte a cartões de crédito, transações,
    orçamentos mensais, categorias e faturas com parcelamento.

    ## Valores Monetários
    Todos os valores monetários são representados como **strings decimais** (ex: `"1234.56"`)
    para preservar precisão e evitar erros de ponto flutuante.

    ## Paginação
    Endpoints de listagem usam **cursor-based pagination**. Parâmetros:
    - `limit` (int, default: 20, max: 100)
    - `cursor` (string, opcional): valor de `pagination.next_cursor` da resposta anterior

    ## Autenticação
    Endpoints protegidos requerem o header `Authorization: Bearer {token}`.
    Obtenha o token em `POST /api/v1/token`.
  license:
    name: MIT
  title: Financial API
  version: 1.0.0
paths:
...
  /api/v1/invoice-items:
    post:
      consumes:
      - application/json
      description: |-
        Registra uma nova compra no cartão de crédito com suporte a parcelamento.
        Para cada parcela é criado um `InvoiceItem` na fatura correspondente ao mês de vencimento.

        **Campos:**
        - `card_id`: UUID do cartão
        - `purchase_date`: data da compra em `YYYY-MM-DD`
        - `total_amount`: valor total da compra (ex: `"1200.00"`)
        - `currency`: `BRL` | `USD` | `EUR` (opcional, default `BRL`)
        - `installment_total`: `1` para à vista, até `48` para parcelado
      parameters:
      - description: Dados da compra
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.PurchaseCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Itens de fatura criados (um por parcela)
          schema:
            $ref: '#/definitions/dtos.PurchaseCreateOutput'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Cartão ou categoria não encontrados
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "503":
          description: Serviço de cartão indisponível
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Criar compra (invoice item)
      tags:
      - invoices
  /api/v1/invoice-items/{id}:
    delete:
      description: |-
        Remove uma compra e todas as suas parcelas das faturas. Esta operação é irreversível.
        O parâmetro `category_id` é opcional e usado para atualizar o orçamento após a remoção.
      parameters:
      - description: ID do item de fatura
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: ID da categoria (para atualizar orçamento)
        format: uuid
        in: query
        name: category_id
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Compra removida com sucesso
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Item não encontrado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Remover compra (invoice item)
      tags:
      - invoices
    put:
      consumes:
      - application/json
      description: |-
        Atualiza os dados de uma compra. A atualização é propagada para todas as parcelas vinculadas.
        O `id` é o ID de qualquer parcela da compra original.
      parameters:
      - description: ID do item de fatura
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados da compra
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.PurchaseUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: Itens atualizados (todas as parcelas)
          schema:
            $ref: '#/definitions/dtos.PurchaseUpdateOutput'
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Item não encontrado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Atualizar compra (invoice item)
      tags:
      - invoices
  /api/v1/invoices:
    get:
      description: |-
        Retorna a lista paginada de faturas filtradas por mês ou por cartão.
        **Exatamente um** dos parâmetros `month` ou `cardId` deve ser informado.

        - **`?month=YYYY-MM`**: lista todas as faturas do usuário no mês especificado (ex: `?month=2025-01`)
        - **`?cardId=uuid`**: lista todas as faturas de um cartão específico (ex: `?cardId=abc-123`)
      parameters:
      - description: Mês de referência (YYYY-MM). Mutuamente exclusivo com cardId.
        example: 2025-01
        in: query
        name: month
        type: string
      - description: ID do cartão (UUID). Mutuamente exclusivo com month.
        format: uuid
        in: query
        name: cardId
        type: string
      - default: 20
        description: 'Itens por página (default: 20, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Cursor de paginação
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista paginada de faturas
          schema:
            $ref: '#/definitions/dtos.InvoicePaginatedOutput'
        "400":
          description: Parâmetros ausentes ou inválidos
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Listar faturas
      tags:
      - invoices
  /api/v1/invoices/{id}:
    get:
      description: |-
        Retorna os detalhes completos de uma fatura, incluindo todos os itens com valores por parcela,
        rótulo de parcelamento (`installment_label`: ex: `"3/12"` ou `"À vista"`), data de vencimento e total.
      parameters:
      - description: ID da fatura
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dados completos da fatura
          schema:
            $ref: '#/definitions/dtos.InvoiceOutput'
        "401":
          description: Não autenticado
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "404":
          description: Fatura não encontrada
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/httperrors.ProblemDetail'
      security:
      - BearerAuth: []
      summary: Buscar fatura por ID
      tags:
      - invoices
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT no formato: `Bearer {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"